feat: new ui
- **SCRAPED** the entire UI.
- Created `doc.ejs` for documentation and blog.
- Created `docs.ejs` for listing all documentations and blogs.
- Created `explore.ejs` for exploring notes with sorting and filtering options.
- Added `codemirror.ejs` for testing CodeMirror editor integration.
- Enhanced styles and scripts for better user experience across all new views.

+++ public/js/modules/date-utils.mjs
@@ -14,7 +14,7 @@ export function absoluteTime(date) {
 export function relativeTime(date) {
 	const now = new Date().getTime();
 
-	let option = { style: 'long', numeric: 'always' };
+	let option = { style: 'short', numeric: 'always' };
 	let args = [];
 	let timeDifference = date - now;
